requirements.txt
appnope==0.1.0
attrs==19.3.0
aws-wsgi==0.2.6
backcall==0.2.0
cazoo-logger==0.8.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
coverage==5.1
cryptography==2.9.2
decorator==4.4.2
docutils==0.15.2
ecdsa==0.15
Flask==1.1.2
Flask-AWSCognito==1.3
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-WTF==0.14.3
idna==2.9
importlib-metadata==1.7.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-mock==3.1.1
pytest-random-order==1.0.4
pytest-testmon==1.0.2
python-dateutil==2.8.1
python-jose==3.1.0
python-util==1.2
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
